chore: revies fixes
